Added Close Event capture

+++ Dynamics365 Customizing Downloader/DownloadMultiple.xaml.cs
@@ -231,10 +231,7 @@ namespace Dynamics365CustomizingDownloader
         /// <param name="e">The <see cref="System.Windows.RoutedEventArgs"/> instance containing the event data.</param>
         private void Btn_close_Click(object sender, RoutedEventArgs e)
         {
-            if (this.downloadIndex == 0)
-            {
-                this.Close();
-            }
+            this.DownloadWindow_Closing(null, null);
         }
 
         /// <summary>
@@ -282,30 +279,37 @@ namespace Dynamics365CustomizingDownloader
 
                 foreach (Xrm.CrmSolution solution in this.CRMSolutions)
                 {
-                    using (Xrm.ToolingConnector toolingConnector = new Xrm.ToolingConnector())
+                    if (!this.worker.CancellationPending)
                     {
-                        // Delete Solution File if it exists
-                        if (File.Exists(Path.Combine(this.selectedPath, solution.Name + ".zip")))
+                        using (Xrm.ToolingConnector toolingConnector = new Xrm.ToolingConnector())
                         {
+                            // Delete Solution File if it exists
+                            if (File.Exists(Path.Combine(this.selectedPath, solution.Name + ".zip")))
+                            {
+                                File.Delete(Path.Combine(this.selectedPath, solution.Name + ".zip"));
+                            }
+
+                            toolingConnector.DownloadSolution(toolingConnector.GetCrmServiceClient(connectionString: this.CRMConnection.ConnectionString), solution.Name, this.selectedPath);
+
+                            Xrm.CrmSolutionPackager crmSolutionPackager = new Xrm.CrmSolutionPackager();
+
+                            if (Directory.Exists(Path.Combine(this.selectedPath, solution.Name)))
+                            {
+                                Directory.Delete(Path.Combine(this.selectedPath, solution.Name), true);
+                                LogToUI($"Delete {Path.Combine(this.selectedPath, solution.Name).ToString()}", true);
+                            }
+
+                            crmSolutionPackager.ExtractCustomizing(Path.Combine(this.selectedPath, solution.Name + ".zip"), Path.Combine(this.selectedPath, solution.UniqueName));
+
                             File.Delete(Path.Combine(this.selectedPath, solution.Name + ".zip"));
+                            LogToUI($"Delete {Path.Combine(this.selectedPath, solution.Name + ".zip").ToString()}", true);
+
+                            this.downloadIndex--;
                         }
-
-                        toolingConnector.DownloadSolution(toolingConnector.GetCrmServiceClient(connectionString: this.CRMConnection.ConnectionString), solution.Name, this.selectedPath);
-
-                        Xrm.CrmSolutionPackager crmSolutionPackager = new Xrm.CrmSolutionPackager();
-
-                        if (Directory.Exists(Path.Combine(this.selectedPath, solution.Name)))
-                        {
-                            Directory.Delete(Path.Combine(this.selectedPath, solution.Name), true);
-                            LogToUI($"Delete {Path.Combine(this.selectedPath, solution.Name).ToString()}", true);
-                        }
-
-                        crmSolutionPackager.ExtractCustomizing(Path.Combine(this.selectedPath, solution.Name + ".zip"), Path.Combine(this.selectedPath, solution.UniqueName));
-
-                        File.Delete(Path.Combine(this.selectedPath, solution.Name + ".zip"));
-                        LogToUI($"Delete {Path.Combine(this.selectedPath, solution.Name + ".zip").ToString()}", true);
-
-                        this.downloadIndex--;
+                    }
+                    else
+                    {
+                        e.Cancel = true;
                     }
                 }
             }
@@ -327,6 +331,28 @@ namespace Dynamics365CustomizingDownloader
             DownloadMultiple.LogToUI("---------------");
             DownloadMultiple.LogToUI("Finished download/extraction");
             DownloadMultiple.LogToUI("---------------");
+        }
+
+        /// <summary>
+        /// Form Closing Event
+        /// </summary>
+        /// <param name="sender">The sender.</param>
+        /// <param name="e">The <see cref="DoWorkEventArgs"/> instance containing the event data.</param>
+        private void DownloadWindow_Closing(object sender, CancelEventArgs e)
+        {
+            if (this.downloadIndex == 0)
+            {
+                this.Close();
+            }
+            else
+            {
+                MessageBoxResult dialogResult = MessageBox.Show("Download is still running, are you sure to abort the process?", "Background thread is still active!", MessageBoxButton.YesNo, MessageBoxImage.Warning, MessageBoxResult.No);
+
+                if (dialogResult == MessageBoxResult.Yes)
+                {
+                    this.worker.CancelAsync();
+                }
+            }
         }
     }
 }
